backend/requirements.txt
asgiref==3.7.2
black==23.11.0
click==8.1.7
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.11.2
django-environ
django-filter==23.4
djangorestframework==3.14.0
geographiclib==2.0
djangorestframework-gis
geopy==2.4.1
inflection==0.5.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.11.2
platformdirs==4.0.0
psycopg2-binary==2.9.9
pytz==2023.3.post1
PyYAML==6.0.1
Pillow
sqlparse==0.4.4
tomli==2.0.1
typing_extensions==4.8.0
uritemplate==4.1.1

